+++ tests/calculator_tests.py
@@ -3,7 +3,7 @@ import unittest
 from wdc.calculator import calc_workday_end
 
 
-class CalcualtorFixture(unittest.TestCase):
+class CalculatorFixture(unittest.TestCase):
     def test_valid_scenarios(self):
         scenarios = [
             ('0800', 30, '0800', '1630'),