+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 ### Add
+- ```export``` command
+
+### Fixed
+- Tasks with no history do not cause error on ```info``` command
+
+### Changed
+- Moved documentation to the wiki
+
+## [0.3.9]
+### Add
 - ```amend``` command
 
 ### Changed